Use background script to determine page ready state
Previously a MutationObserver was used, however, this has the problem,
that when navigating from the YouTube home page to the watch later list
the content script would not be executed.

Using a background script to trigger the content script works much
better, however, it sadly still comes with one challenge that needs to
be solved. When reloading the watch later list no remove buttons are
created.