Improve page ready detection
By watching the status attribute of a tag reloads on the watch later
list now work as intended. However, there is a slight but still present
to this solution. When leaving the watch later list an additional remove
button is created behind the scenes.

However, in this case this is not an issue because the remove buttons
need to be destroyed anyway when navigating to the watch later list to
prevent duplicate remove buttons on the watch later page.