Implement iteration limit for IntervalRunner
Currently, many background scripts do not handle the case where for a given
NavigationFilter no element could be found. This would cause IntervalRunner to
continue running until the browser tab gets closed.

To prevent this a default iteration limit is now present. This iteration limit
will cause the currently running interval to automatically stop when a given
number of iterations have occurred in the interval.