Extend HtmlTreeDirectNavigator to allow specification of HTML tree fragments
The intention of the HtmlTreeDirectNavigator class is to make the search for
HTML elements easy by not having to provide an entire tree of HTML elements.
However, sometimes similar target elements exist under a different HTML tree
causing the HtmlTreeDirectNavigator to return those elements which is not
desirable.

To address this issue this commit extends the HtmlTreeDirectNavigator with a
`filter` function. Each filter call on the HtmlTreeDirectNavigator must be found
 in the HTML tree. If any filter does not match with an element in the HTMl, the
  final result will be an empty array (in the case of `find`).