Remove HtmlTreeNavigator
The HtmlTreeNavigator is a bit clunky to use. A developer needs to model the
entire HTML tree to find a single target element, which becomes an annoyance
when done many times.

Due to this annoyance the HtmlTreeDirectNavigator was created which does not
require to specify all elements in HTML tree. In the next commit the
HtmlTreeDirectNavigator will be renamed to HtmlTreeNavigator