Use SVG draw paths or assumptions to trigger YouTube actions
This commit changes the content scripts to use other means for determining the
right actions. Instead of using text inside HTML elements, now SVG paths are
used which fixes the extension not working for all languages except English.

For the `video.ts` script it was not possible to select the watch later checkbox
in the playlists choice popup via an SVG path. Instead, always the first
element is checked since YouTube places the watch later playlist as the first
entry.