Initialize drop-down menu for most content scripts programmatically
So far it has been assumed that the order in which drop-down popups are
initialized is consistent. The first time the drop-down is opened all content is
loaded (HTML is added or removed). Then on consequent clicks elements get the
hidden attribute assigned or removed.
However, this order is not the case when triggering quick actions and navigating
around different pages. When going from the "Watch Later" playlist into the
playing "Watch Later" playlist the first click on the popup drop-down causes
the previous content to be hidden first and then removed or added in the same
first click.

This lack of consistency makes it hard to process the changes with
mutation observers or mutation summaries because now some form of state is
required to know if a drop-down has already been initialized. This is quite
trivial for scripts like playlist.ts, home-page.ts and watching-playlist.ts.
However, for video.ts this state becomes harder to manage because the element
which triggers the "Save To" dialog can be outside a drop-down or inside a
drop-down.

Therefore, to keep things simple enough the all content scripts except video.ts
now programmatically open and close the more option button drop-downs. The
reason video.ts is not included is that clicking the more options button
programmatically twice does not seem to actually do anything. So there is a
chance for the extension to fail on video.ts in complex cases.

In the future, it will be worth looking deeper into a solution that truly is
consistent for mutation summaries or mutation observers. But for now, this shall
do.