Use RxJs to optimize the delay for video.ts quick action
Because the pattern of the changes for YouTube's popup container is different
when navigating between pages, some form of state is required to get a complete
list of all mutation summary changes.

The easiest solution would be to store all changes in a list and implement a
delay, however, that raises the following questions: Is it safe to assume that
the delay will be consistent on every machine of every user? How long should the
delay be?

Additionally, we also don't want a too long delay because that does affect the
length of the operation we want ot perform.

With RxJs the above topics can be mostly resolved because of the operators it
offers to deal with async sources.