Remove dynamic url setting in chrome manifest
This setting was added to the V3 manifest without being functionally implemented until Chrome version 130. Dynamic urls can only be used with the `chrome.runtime.getURL()` API. If this setting is used without accessing the resource over `chrome.runtime.getURL()`, the extension is now using unaccessible resources.

Note: By dropping this setting, it's possible to fingerprint users by checking if the resources provided by the extension can be accessed. Since this behavior would be the case with Chrome versions before 130 anyway, it's not too much of a concern.